Weight Completion Service/requirements.txt
# Works with python 3.6.8

# Trouble?
# Install numpy manually (still this version) instead of via requirements
# Install pyemd manually (still this versio)
# Make sure you have the C++ build tools (14.0.0 or higher) -> maybe check the C++/CLI support for v142 build tools (14.24)

alembic==1.4.3
SQLAlchemy==1.3.20
Flask==1.1.2
Flask_Cors==3.0.9
Flask_SQLAlchemy==2.4.4
flask_migrate==2.5.3
flask_script==2.0.6
flask_serialize==1.4.0
geopy==1.23.0
networkx==2.1
numpy==1.16.6
pandas==0.20.3
# pandas==1.1.4
pyemd==0.5.1
scikit_learn==0.23.2
scipy==1.5.4
tqdm==4.53.0
tensorflow==1.14.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
psycopg2-binary==2.8.6
python-dotenv~=0.15.0
#pydriosm
python-dateutil~=2.8.1
scikit-learn~=0.23.2